Upgrade build to VS 2017, MSBuild 15
