+++ Test/CoreSDK.Test/Shared/Core.Shared.Tests.shproj
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Label="Globals">
     <ProjectGuid>c4a0db5b-0988-4e04-b9b8-bcf3b0842000</ProjectGuid>
   </PropertyGroup>
