+++ Test/ServerTelemetryChannel.Test/Shared.Tests/Shared.Tests.shproj
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Label="Globals">
     <ProjectGuid>C5A3F18F-1426-4941-9D81-D21062A2B8F2</ProjectGuid>
   </PropertyGroup>
