--- src/Core/Managed/PCL/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Desktop.Analyzers" version="1.1.0" targetFramework="portable45-net45+win8" />
-  <package id="MicroBuild.Core" version="0.2.0" targetFramework="portable45-net45+win8" developmentDependency="true" />
-  <package id="StyleCop.MSBuild" version="4.7.55.0" targetFramework="portable45-net45+win8" developmentDependency="true" />
-</packages>
