+++ src/Core/Managed/Product.props
@@ -6,6 +6,7 @@
     <CorePath>$(RelativeOutputPathBase.Substring(4))</CorePath>	
     <OutputPath>$(BinRoot)\$(Configuration)\$(CorePath)</OutputPath>
     <OutputPath>$([System.IO.Path]::GetFullPath( $(OutputPath) ))\</OutputPath>
+    <BaseIntermediateOutputPath>$(BaseIntermediateOutputPath)\$(MSBuildProjectName)</BaseIntermediateOutputPath>
   </PropertyGroup>
   
   <PropertyGroup>
