+++ src/Core/Managed/Shared/Extensibility/Implementation/JsonWriter.cs
@@ -89,11 +89,7 @@
         {
             if (value.HasValue)
             {
-#if NET45 || NET46
                 this.WriteProperty(name, value.Value.ToString(string.Empty, CultureInfo.InvariantCulture));
-#else
-                this.WriteProperty(name, value.Value.ToString(CultureInfo.InvariantCulture, string.Empty));
-#endif
             }
         }
