+++ src/Core/Managed/Shared/Utils.cs
@@ -61,11 +61,7 @@
         public static TimeSpan ValidateDuration(string value)
         {
             TimeSpan interval;
-#if NET45 || NET46
             if (!TimeSpan.TryParse(value, CultureInfo.InvariantCulture, out interval))
-#else
-            if (!TimeSpanEx.TryParse(value, CultureInfo.InvariantCulture, out interval))
-#endif
             {
                 CoreEventSource.Log.TelemetryIncorrectDuration();
                 return TimeSpan.Zero;
