+++ src/Core/NuGet/Core/Package.nuspec
@@ -29,33 +29,14 @@
         <dependency id="Microsoft.Bcl.Async" version="1.0.168" />
       </group>
       <group targetFramework="netstandard1.3">
-        <dependency id="System.Collections" version="4.0.11" />
-        <dependency id="System.Diagnostics.Debug" version="4.0.11" />
-        <dependency id="System.Diagnostics.Tools" version="4.0.1" />
-        <dependency id="System.Diagnostics.Tracing" version="4.1.0" />
-        <dependency id="System.Globalization" version="4.0.11" />
-        <dependency id="System.IO" version="4.1.0" />
-        <dependency id="System.IO.Compression" version="4.1.0" />
-        <dependency id="System.Linq" version="4.1.0" />
-        <dependency id="System.Net.Http" version="4.1.0" />
-        <dependency id="System.Reflection" version="4.1.0" />
-        <dependency id="System.Reflection.Extensions" version="4.0.1" />
-        <dependency id="System.Runtime" version="4.1.0" />
-        <dependency id="System.Runtime.Extensions" version="4.1.0" />
-        <dependency id="System.Text.Encoding" version="4.0.11" />
-        <dependency id="System.Text.Encoding.Extensions" version="4.0.11" />
-        <dependency id="System.Text.RegularExpressions" version="4.1.0" />
-        <dependency id="System.Threading" version="4.0.11" />
-        <dependency id="System.Threading.Tasks" version="4.0.11" />
-        <dependency id="System.Xml.XDocument" version="4.0.11" />
       </group>
     </dependencies>
   </metadata>
   <files>
-    <!-- Binaries for .NET PCL -->
-    <file src="$configuration$\Core\Managed\PCL\Microsoft.ApplicationInsights.dll" target="lib\netstandard1.3" />
-    <file src="$configuration$\Core\Managed\PCL\Microsoft.ApplicationInsights.pdb" target="lib\netstandard1.3" />
-    <file src="$configuration$\Core\Managed\PCL\Microsoft.ApplicationInsights.xml" target="lib\netstandard1.3" />
+    <!-- Binaries for .NET Standard 1.3 -->
+    <file src="$configuration$\Core\Managed\NetStandard13\netstandard1.3\Microsoft.ApplicationInsights.dll" target="lib\netstandard1.3" />
+    <file src="$configuration$\Core\Managed\NetStandard13\netstandard1.3\Microsoft.ApplicationInsights.pdb" target="lib\netstandard1.3" />
+    <file src="$configuration$\Core\Managed\NetStandard13\netstandard1.3\Microsoft.ApplicationInsights.xml" target="lib\netstandard1.3" />
 
     <!-- Binaries for .NET 4.0 projects -->
     <file src="$configuration$\Core\Managed\Net40\Microsoft.ApplicationInsights.dll" target="lib\net40" />
