+++ src/TelemetryChannels/ServerTelemetryChannel/Shared/Shared.shproj
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup Label="Globals">
     <ProjectGuid>{D03AC224-D6C9-4C64-9B82-619E28FF9AC6}</ProjectGuid>
   </PropertyGroup>
